third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-2/mixed-case-color-aware-properties.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/elements-test.js"></script>
<script>

function test() {
  var colorAwareProperties = ['bAckground-ColoR', 'COloR', 'Border-coLoR', 'border-right-color', 'BOX-SHADOW'];
  for (var i = 0; i < colorAwareProperties.length; ++i) {
    var isColorAware = SDK.cssMetadata().isColorAwareProperty(colorAwareProperties[i]);
    TestRunner.addResult(colorAwareProperties[i] + (isColorAware ? ' is' : ' is NOT') + ' color aware');
  }
  TestRunner.completeTest();
}
</script>
</head>

<body onload="runTest()">
<p>
Tests that color-related mix-cased CSS properties are actually color aware.
</p>
</body>
</html>
