third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-2/mixed-case-color-aware-properties-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that color-related mix-cased CSS properties are actually color aware.

bAckground-ColoR is color aware
COloR is color aware
Border-coLoR is color aware
border-right-color is color aware
BOX-SHADOW is color aware
